board: sifive: Fix -Wint-to-pointer-cast warning
The following warning is seen in unleashed.c in a 32-bit build:

  warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]

Cast with uintptr_t.

+++ board/sifive/unmatched/unmatched.c
@@ -15,7 +15,7 @@ void *board_fdt_blob_setup(void)
 {
 	if (IS_ENABLED(CONFIG_OF_SEPARATE)) {
 		if (gd->arch.firmware_fdt_addr)
-			return (ulong *)gd->arch.firmware_fdt_addr;
+			return (ulong *)(uintptr_t)gd->arch.firmware_fdt_addr;
 	}
 
 	return (ulong *)&_end;
